Normalized module config
