+++ src/utils.js
@@ -1,3 +1,5 @@
-export const NAMESPACES = ['collection', 'member']
+const NAMESPACES = ['collection', 'member']
+
+export const reduceNamespaces = (reducer) => NAMESPACES.reduce(reducer, {})
 
 export const forOwn = (obj, iteratee) => Object.keys(obj).forEach((key) => iteratee(obj[key], key))
